Changing variables back loop to integers (ushort has overflown and caused infinite loops)
--HG--
branch : dev

+++ Samples/Hast.Samples.SampleAssembly/ImageContrastModifier.cs
@@ -33,7 +33,7 @@ namespace Hast.Samples.SampleAssembly
 
             contrastValue = (100.0 + contrastValue) / 100.0;
 
-            for (ushort i = 0; i < imageHeight * imageWidth * 3; i++)
+            for (int i = 0; i < imageHeight * imageWidth * 3; i++)
             {
                 pixel = (ushort)memory.ReadUInt32(i + ChangeContrast_ImageStartIndex);
                 memory.WriteUInt32(i + ChangeContrast_ImageStartIndex, ChangePixelValue(pixel, contrastValue));
